app-emulation/dxvk-winelib: <PLEASE READ NOTE>
I've been having issues building dxvk-winelib-9999 which is related to
the recent commit related to dxgi. I have not found the cause, but it
might have to do with wine version, because the new dxgi stuff
requires GCC 8 to build, so this might be an issue related to wineg++.

I have not used the winelib version of DXVK for some time, so I've
been thinking about dropping this ebuild all together, or at least
dropping the 9999 version.

Package-Manager: Portage-3.0.8, Repoman-3.0.2