app-emulation/dxvk-winelib: As I suspected
This refers to the previous commit:

Looks like it was the wine version. My current wine version is 5.7 which
is built from lutris's wine fork. Switching to 5.14 which I had
installed for testing purposes was able to build dxvk-winelib-9999.

I'm not entirely sure what the earliest version of wine is required to
build dxvk-winelib with the new dxgi stuff, but I'll update this
ebuild's DEPENDS when I find out.