gui-wm/hyprland: Add live version and eglstreams
Enabling the eglstreams useflag replaces the provided wlroots with
wlroots-eglstreams which is a fork that adds eglstreams support for
Nvidia users.

As an Nvidia user I find that using the normal GBM to be pretty usable,
but has some flickering issues. I'm just curious to see if
wlroots-eglstreams makes any improvements.